Move test pages to gallery
