+++ Xamarin.Forms.Controls/GalleryPages/PlatformSpecificsGalleries/VisualElementiOS.cs
@@ -1,28 +1,27 @@
-﻿using Xamarin.Forms.CustomAttributes;
-using Xamarin.Forms.Internals;
+using System.Windows.Input;
 using Xamarin.Forms.PlatformConfiguration;
 using Xamarin.Forms.PlatformConfiguration.iOSSpecific;
 
-namespace Xamarin.Forms.Controls
+namespace Xamarin.Forms.Controls.GalleryPages.PlatformSpecificsGalleries
 {
-	[Preserve(AllMembers = true)]
-	[Issue(IssueTracker.Bugzilla, 90000, "Platform Specifics - iOS Blur Effect")]
-	public class PlatformSpecifics_iOSBlurEffect : TestContentPage
+	public class VisualElementiOS : ContentPage
 	{
-		protected override void Init()
+		public VisualElementiOS(ICommand restore)
 		{
+			var restoreButton = new Button { Text = "Back To Gallery" };
+			restoreButton.Clicked += (sender, args) => restore.Execute(null);
+
 			var image = new Image { Source = ImageSource.FromFile("crimson.jpg") };
 			var box = new BoxView { HeightRequest = 300, WidthRequest = 600 };
 			box.On<iOS>().UseBlurEffect(BlurEffectStyle.Light);
-
 			Button button1 = GetButton(box, BlurEffectStyle.None);
 			Button button2 = GetButton(box, BlurEffectStyle.ExtraLight);
 			Button button3 = GetButton(box, BlurEffectStyle.Light);
 			Button button4 = GetButton(box, BlurEffectStyle.Dark);
-
-			var buttons = new StackLayout { Orientation = StackOrientation.Horizontal, Children = { button1, button2, button3, button4 }};
+			var buttons = new StackLayout { Orientation = StackOrientation.Horizontal, Children = { button1, button2, button3, button4 } };
 
 			Content = new StackLayout { Children = { buttons, new AbsoluteLayout { Children = { image, box } } } };
+			Title = "Visual Element Features";
 		}
 
 		Button GetButton(BoxView box, BlurEffectStyle value)
